Move transactional tests into their own file
